+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 ## Release Notes
 
-### Bugfixes
+### Changes
 
 **Tables Management**
 
@@ -13,4 +13,9 @@
 **Table Management: PinUP Popper Table Data**
 
 - Added "Game Type", "Game Version (File Version)" and "Game Theme" fields.
-- 
+- Added "Auto-Fill" button which fills missing data by looking them up from the table metadata and VPS.
+
+
+### Bugfixes
+
+- Removed table data field "ROM Url" (I have no idea why I added this, it was stored nowhere).